feat: Add isLoading state to AuthContext
This commit adds the isLoading state to the AuthContext in order to track the loading status of authentication requests. The isLoading state is set to true when a login or logout request is initiated, and set to false when the request is completed. This state will be used to display loading spinners or disable UI elements during authentication requests.